move the display driver onto main()
